Coding Days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2663 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/basis/modpoly/ModifiedPolyBasis.hpp
@@ -109,7 +109,7 @@ private:
 		double x_val = x_store[0] + val * pow(2.0, -(1.0)*(static_cast<double>(deg)));
 
 		//Horner
-		for(int i = deg-2; i >= 0; i--)
+		for(size_t i = deg-2; i >= 0; i--)
 		{
 			y_val = y_val * x_val + y_store[i];
 		}
